src/test/java/numbers/RationalTest.java
package numbers;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;
import junit.framework.Test;
import junit.framework.TestCase;
import junit.framework.TestSuite;

/**
 * Unit test for simple App.
 */
public class RationalTest 
    extends TestCase
{
    /**
     * Create the test case
     *
     * @param testName name of the test case
     */
    public RationalTest( String testName )
    {
        super( testName );
    }

    /**
     * @return the suite of tests being tested
     */
    public static Test suite()
    {
        return new TestSuite( RationalTest.class );
    }



    /**
     * Rigourous Test :-)
     */
    public void testApp()
    {
        // Given no preconditions

        // When I create a default `Rational` value
        Rational value = new Rational();
        // Then the value should have numerator 0
        assertThat("the numerator should be 0", value.numerator(), is(0));
        // And the value should have denominator 1
        assertThat("the denominator should be 1", value.denominator(), is(1));

        Rational value2 = new Rational(2);
        // Then the value should have numerator `2`
        assertThat("the numerator should be 2", value2.numerator(), is(2));
        // And the value should have denominator `1`
        assertThat("the denominator should be 1", value2.denominator(), is(1));
    }
}
